variation graph code reorganization
